homework/hw БЕЗОПАСНОСТЬ/task7/sql-injection-lab/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
bcrypt
pydantic
pytest
httpx
pytest-asyncio
asgi-lifespan
sqlalchemy[asyncio]>=2.0.25
alembic